refactor: extract dispatch security preflight
